Move create study logic and draft import controller.
Moving create study so that any code in the future can just call the study service with a study to create the study.
You can still invoke the connection based version if you need to create a study as part of a larger transaction, but
we should document that steps taken in createStudy without a connection must be taken.